crewai_service/requirements.txt
flask
crewai
python-dotenv

# If CrewAI uses OpenAI under the hood, you may also need:
# openai
flask-cors
